Added Google API methods to service
